Added MainCode, a function in extrafunctions, a function in scraping file
A folder named CollectData now contains a main driver code which uses the functions from modules folder. Another Function in Scraping.py was added which utilises the two functions already present in it to scrape the data. This function also uses another function that i added in extraFunctions,py which appends the data scraped into the CSV file which will be found in the Scraped Folder named as ScrapedData.csv